♻️ refactor the scenario folder
