+++ tactics2d/envs/parking.py
@@ -11,9 +11,9 @@ from tactics2d.map.element import Map
 from tactics2d.map.generator import ParkingLotGenerator
 from tactics2d.participant.element import Vehicle
 from tactics2d.physics import SingleTrackKinematics
-from tactics2d.scenario import ScenarioManager, RenderManager
-from tactics2d.scenario.traffic_event import TrafficEvent
-from tactics2d.sensor import TopDownCamera, SingleLineLidar
+from tactics2d.traffic import TrafficScenarioManager
+from tactics2d.traffic.violation_detection import TrafficEvent
+from tactics2d.sensor import TopDownCamera, SingleLineLidar, RenderManager
 from tactics2d.trajectory.element import State
 
 from tactics2d.participant.element.defaults import VEHICLE_MODEL
@@ -43,7 +43,7 @@ def truncate_angle(angle: float):
     return angle
 
 
-class ParkingScenarioManager(ScenarioManager):
+class ParkingScenarioManager(TrafficScenarioManager):
     """This class provides a parking scenario manager.
 
     Attributes:
