+++ tactics2d/envs/racing.py
@@ -10,10 +10,10 @@ from gymnasium.error import InvalidAction
 from tactics2d.map.element import Map
 from tactics2d.participant.element import Vehicle
 from tactics2d.trajectory.element import State
-from tactics2d.sensor import TopDownCamera
+from tactics2d.sensor import TopDownCamera, RenderManager
 from tactics2d.map.generator import RacingTrackGenerator
-from tactics2d.scenario import ScenarioManager, RenderManager
-from tactics2d.scenario.traffic_event import TrafficEvent
+from tactics2d.traffic import TrafficScenarioManager
+from tactics2d.traffic.violation_detection import TrafficEvent
 
 
 STATE_W = 200
@@ -39,7 +39,7 @@ DISCRETE_ACTION = np.array(
 THRESHOLD_NON_DRIVABLE = 0.5
 
 
-class RacingScenarioManager(ScenarioManager):
+class RacingScenarioManager(TrafficScenarioManager):
     """_summary_
 
     Attributes:
