--- tactics2d/scenario/__init__.py
@@ -1,4 +1,0 @@
-from .render_manager import RenderManager
-from .scenario_manager import ScenarioManager
-
-__all__ = ["RenderManager", "ScenarioManager"]
