+++ tactics2d/sensor/__init__.py
@@ -1,5 +1,6 @@
 from .sensor_base import SensorBase
 from .camera import TopDownCamera
 from .lidar import SingleLineLidar
+from .render_manager import RenderManager
 
-__all__ = ["SensorBase", "TopDownCamera", "SingleLineLidar"]
+__all__ = ["SensorBase", "TopDownCamera", "SingleLineLidar", "RenderManager"]
