+++ tactics2d/sensor/render_manager.py
