+++ tactics2d/traffic/__init__.py
@@ -1,0 +1,3 @@
+from .traffic_scenario_manager import TrafficScenarioManager
+
+__all__ = ["TrafficScenarioManager"]
