+++ tactics2d/traffic/violation_detection/detect_collision.py
@@ -5,7 +5,7 @@ class CollisionDetector:
     @staticmethod
     def detect_static_collision(trace: Geometry, obstacles: Geometry) -> bool:
         """Detect if the agent collides with the static obstacles."""
-        
+
         return False
 
     @staticmethod
