+++ tactics2d/traffic/violation_detection/traffic_event.py
